+++ test/generated-custom-exports.js
@@ -1,0 +1,12 @@
+module.exports = {
+	customMedia: {
+		'--narrow-window': '(max-width: 30em)',
+		'--dark-mode': '(prefers-color-scheme: dark)'
+	},
+	customProperties: {
+		'--order': '1'
+	},
+	customSelectors: {
+		':--heading': 'h1, h2, h3, h4, h5, h6'
+	}
+};