+++ test/generated-custom-exports.mjs
@@ -1,0 +1,12 @@
+export const customMedia = {
+	'--narrow-window': '(max-width: 30em)',
+	'--dark-mode': '(prefers-color-scheme: dark)'
+};
+
+export const customProperties = {
+	'--order': '1'
+};
+
+export const customSelectors = {
+	':--heading': 'h1, h2, h3, h4, h5, h6'
+};